refactored code
simplified directory to avoid the com/example/.../springtemplate nested structure, and updated package names
