+++ src/main/java/Role.java
@@ -1,0 +1,10 @@
+package main.java;
+
+public enum Role {
+  CUSTOMER, EMPLOYEE, MANAGER;
+
+  @Override
+  public String toString() {
+    return super.toString();
+  }
+}
