--- src/main/java/com/example/springtemplate/DemoApplication.java
@@ -1,13 +1,0 @@
-package com.example.springtemplate;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
-	}
-
-}
