--- src/main/java/com/example/springtemplate/Role.java
@@ -1,10 +1,0 @@
-package com.example.springtemplate;
-
-public enum Role {
-  CUSTOMER, EMPLOYEE, MANAGER;
-
-  @Override
-  public String toString() {
-    return super.toString();
-  }
-}
