--- src/main/java/com/example/springtemplate/orders/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.example.springtemplate.orders;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface OrderRepository
-        extends CrudRepository<Order, Integer> {
-}
