--- src/main/java/com/example/springtemplate/product_orders/ProductOrderRepository.java
@@ -1,7 +1,0 @@
-package com.example.springtemplate.product_orders;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ProductOrderRepository
-        extends CrudRepository<ProductOrder, Integer> {
-}
