--- src/main/java/com/example/springtemplate/products/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.example.springtemplate.products;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ProductRepository
-        extends CrudRepository<Product, Integer> {
-}
