--- src/main/java/com/example/springtemplate/users/UserRepository.java
@@ -1,7 +1,0 @@
-package com.example.springtemplate.users;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository
-        extends CrudRepository<User, Integer> {
-}
