+++ src/main/java/orders/OrderRepository.java
@@ -1,0 +1,7 @@
+package main.java.orders;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface OrderRepository
+        extends CrudRepository<Order, Integer> {
+}
