+++ src/main/java/product_orders/ProductOrderRepository.java
@@ -1,0 +1,7 @@
+package main.java.product_orders;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface ProductOrderRepository
+        extends CrudRepository<ProductOrder, Integer> {
+}
