+++ src/main/java/products/ProductRepository.java
@@ -1,0 +1,7 @@
+package main.java.products;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface ProductRepository
+        extends CrudRepository<Product, Integer> {
+}
