+++ src/main/java/users/UserRepository.java
@@ -1,0 +1,8 @@
+package main.java.users;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository
+        extends CrudRepository<User, Integer> {
+
+}
