--- src/main/other/sql/db_design/db_design_ratings.sql
@@ -1,61 +1,0 @@
--- MySQL dump 10.13  Distrib 5.7.17, for macos10.12 (x86_64)
---
--- Host: 127.0.0.1    Database: db_design
--- ------------------------------------------------------
--- Server version	5.7.19
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!40101 SET NAMES utf8 */;
-/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
-/*!40103 SET TIME_ZONE='+00:00' */;
-/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
-/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
-/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
-/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
-
---
--- Table structure for table `ratings`
---
-
-DROP TABLE IF EXISTS `ratings`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!40101 SET character_set_client = utf8 */;
-CREATE TABLE `ratings` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `likes` tinyint(1) DEFAULT NULL,
-  `score` int(11) DEFAULT NULL,
-  `comment` varchar(45) DEFAULT NULL,
-  `user_id` int(11) DEFAULT NULL,
-  `post_id` int(11) DEFAULT NULL,
-  `created` datetime DEFAULT CURRENT_TIMESTAMP,
-  `updated` datetime DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`),
-  KEY `ratings_2_user_idx` (`user_id`),
-  KEY `ratings_2_post_idx` (`post_id`),
-  CONSTRAINT `ratings_2_post` FOREIGN KEY (`post_id`) REFERENCES `posts` (`id`) ON DELETE CASCADE ON UPDATE CASCADE,
-  CONSTRAINT `ratings_2_user` FOREIGN KEY (`user_id`) REFERENCES `users` (`id`) ON DELETE CASCADE ON UPDATE CASCADE
-) ENGINE=InnoDB AUTO_INCREMENT=21 DEFAULT CHARSET=latin1;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `ratings`
---
-
-LOCK TABLES `ratings` WRITE;
-/*!40000 ALTER TABLE `ratings` DISABLE KEYS */;
-INSERT INTO `ratings` VALUES (1,1,NULL,NULL,3,1,'2021-01-11 17:54:29','2021-01-11 17:54:29'),(2,0,NULL,NULL,4,2,'2021-01-11 18:07:25','2021-01-11 18:07:25'),(3,1,NULL,NULL,2,1,'2021-01-11 23:12:10','2021-01-13 08:15:25'),(4,1,NULL,NULL,NULL,1,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(5,0,NULL,NULL,1,1,'2021-01-11 23:12:10','2021-01-13 22:26:21'),(6,NULL,3,NULL,NULL,1,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(7,NULL,4,NULL,NULL,1,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(8,NULL,5,NULL,NULL,1,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(9,NULL,5,NULL,NULL,1,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(10,1,NULL,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(11,1,NULL,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(12,1,NULL,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(13,0,NULL,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(14,0,NULL,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(15,1,NULL,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(16,NULL,4,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(17,NULL,4,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(18,NULL,5,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(19,NULL,4,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10'),(20,NULL,3,NULL,NULL,2,'2021-01-11 23:12:10','2021-01-11 23:12:10');
-/*!40000 ALTER TABLE `ratings` ENABLE KEYS */;
-UNLOCK TABLES;
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2021-01-13 22:42:23
