--- src/main/other/sql/querying-relational-databases/alter-table-replies-add-replies-2-replies.sql
@@ -1,9 +1,0 @@
-ALTER TABLE `db_design`.`replies`
-ADD COLUMN `replies_id` INT NULL AFTER `post_id`,
-ADD INDEX `replies_2_replies_idx` (`replies_id` ASC);
-ALTER TABLE `db_design`.`replies`
-ADD CONSTRAINT `replies_2_replies`
-  FOREIGN KEY (`replies_id`)
-  REFERENCES `db_design`.`replies` (`id`)
-  ON DELETE CASCADE
-  ON UPDATE CASCADE;
