--- src/main/other/sql/querying-relational-databases/create-table-topics.sql
@@ -1,3 +1,0 @@
-CREATE TABLE `db_design`.`topics` (
-  `topic` VARCHAR(45) NOT NULL,
-  PRIMARY KEY (`topic`));
