--- src/main/other/sql/querying-relational-databases/drop-users.sql
@@ -1,1 +1,0 @@
-DROP TABLE IF EXISTS `db_design`.`users`;
