--- src/main/other/sql/querying-relational-databases/insert-topics.sql
@@ -1,2 +1,0 @@
-INSERT INTO `db_design`.`topics` (`topic`) VALUES ('SPACE');
-INSERT INTO `db_design`.`topics` (`topic`) VALUES ('CARS');
