+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/retail_store_fp?serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=jiajia2002
+spring.datasource.password=P@ssw0rd
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.hibernate.naming-strategy=org.hibernate.cfg.ImprovedNamingStrategy
