--- src/test/java/com/example/springtemplate/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.springtemplate;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
